Reorganizadas pastas do projeto
